% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cos_similarity}
\alias{cos_similarity}
\title{A function to calculate the pairwise cosine similarity between two matrices}
\usage{
cos_similarity(A, B)
}
\arguments{
\item{A}{A (n*p) user*item rating matrix}

\item{B}{A (m*p) user*item rating matrix}
}
\value{
A (m*n) matrix, where (i,j) element representing the cosine similarity 
between row vector A_j and row vector B_i
}
\description{
A function to calculate the pairwise cosine similarity between two matrices
}
