% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{funk_predict}
\alias{funk_predict}
\title{A function to predict ratings of given users and items from Funk SVD factors}
\usage{
funk_predict(U, V, users, items)
}
\arguments{
\item{U}{A (n*k) matrix of n user factors, output from funk_svd()}

\item{V}{A (p*k) matrix of p item factors, output from funk_svd()}

\item{users}{A u length vector that contains zero-based indices of selected users}

\item{items}{A i length vector taht contains zero-based indices of selected items}
}
\value{
A (u*i) matrix with predicted ratings of i items from u users
}
\description{
A function to predict ratings of given users and items from Funk SVD factors
}
\examples{
# Example code for Funk SVD
data("MovieLenseSub") # A subset of MovieLense Data
train <- MovieLenseSub[1:200,1:500]
train[is.na(train)]<-0

res<-funk_svd(train,max_iter=2000)

users<-c(1,2,3)
items<-c(4,5,6,7)

funk_predict(res$U,res$V,users-1,items-1)
}
