% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{funk_svd}
\alias{funk_svd}
\title{A function to decompose rating matrix with missing values through stochastic gradient descent}
\usage{
funk_svd(X, n_features = 10L, alpha = 0.001, gamma = 0.015, tol = 1e-06,
  max_iter = 200L)
}
\arguments{
\item{X}{A (n*p) user by item rating matrix. Missing ratings should be represented
as 0. Non-missing ratings should be re-scale to exclude zero.}

\item{n_features}{Dimension of each factor vector}

\item{alpha}{Learning rate}

\item{gamma}{Regularization parameter}

\item{tol}{Tolerance for convergence checking}

\item{max_iter}{Maximum iterations}
}
\value{
A list of following outputs:
\item{U}{A (n*k) matrix of n user factors}
\item{V}{A (p*k) matrix of p item factors}
\item{Iterations}{Numer of iterations operated}  
\item{Errors}{Root mean square error (RMSE) of the last iteration}
}
\description{
A function to decompose rating matrix with missing values through stochastic gradient descent
}
\examples{
# Example code for Funk SVD
data("MovieLenseSub") # A subset of MovieLense Data
train <- MovieLenseSub[1:200,1:500]
train[is.na(train)]<-0

res<-funk_svd(train,max_iter=2000)

users<-c(1,2,3)
items<-c(4,5,6,7)

funk_predict(res$U,res$V,users-1,items-1)
}
