% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubcf_r.R
\name{get_neighbors}
\alias{get_neighbors}
\title{A function to get the index of the k nearest neighbors}
\usage{
get_neighbors(sim, k)
}
\arguments{
\item{sim}{A (m*n) matrix that contains the pairwise similarity 
between m testing users and n training users, output of function cos_similarity()}

\item{k}{An integer, the number of nearest neighbors to be returned}
}
\value{
A (m*k) matrix, each row contains the index of 
k nearest neighbors of testing users
}
\description{
A function to get the index of the k nearest neighbors
}
