% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize}
\alias{normalize}
\title{A function to remove individual rating bias from each user}
\usage{
normalize(train, test)
}
\arguments{
\item{train}{A (n*p) user*item rating matrix from training users}

\item{test}{A (m*p) user*item rating matrix from testing users}
}
\value{
A list of following outputs:
\item{train_norm}{A (n*p) normalized rating matrix for training users}
\item{test_norm}{A (m*p) normalized rating matrix for testing users}
\item{test_Center}{A m length vector with the mean rating of each testing user}
}
\description{
A function to remove individual rating bias from each user
}
