% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_u}
\alias{predict_u}
\title{A function to predict (normalized) ratings given k nearest neighbors of testing users}
\usage{
predict_u(sim, neighbors, train)
}
\arguments{
\item{sim}{A (m*n) matrix that contains the pairwise similarity 
between m testing users and n training users, output of function cos_similarity()}

\item{neighbors}{A (m*k) matrix, each row contains the index of 
k nearest neighbors of testing users, output of function get_neighbors()}

\item{train}{A (n*p) rating matrix from n users of training set}
}
\value{
A (m*p) matrix of predicted ratings for m testing users on all items
}
\description{
A function to predict (normalized) ratings given k nearest neighbors of testing users
}
