% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubcf_r.R
\name{top_n_list}
\alias{top_n_list}
\title{A function to return Top N recommendations for testing users}
\usage{
top_n_list(test, pred, N = 10)
}
\arguments{
\item{test}{A (m*p) original rating matrix from testing users}

\item{pred}{A (m*p) predicted rating matrix from testing users, 
output of function ubcf_predict()}

\item{N}{An integer, top N items to be recommended}
}
\value{
A (m*N) matrix, each row is the index of recommended N items for the user
}
\description{
A function to return Top N recommendations for testing users
}
\examples{
# Example code for UBCF
data("MovieLenseSub") # A subset of MovieLense Data
train <- MovieLenseSub[1:200,]
test <- MovieLenseSub[201:250,]
train[is.na(train)]<-0
test[is.na(test)]<-0

predicted<-ubcf_predict(train,test,k_neighbors=25)
topN.idx<-top_n_list(test,predicted,N=10)
}
