% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubcf_r.R
\name{ubcf_predict}
\alias{ubcf_predict}
\title{A function to predict ratings for testing users, given training user rating matrix.}
\usage{
ubcf_predict(train, test, k_neighbors)
}
\arguments{
\item{train}{A (n*p) user*item rating matrix from training users. 
Missing ratings should be represented as 0. 
Non-missing ratings should be re-scale to exclude zero.}

\item{test}{A (m*p) user*item rating matrix from testing users. 
Representation should be the same as training data.}

\item{k_neighbors}{An integer, the number of nearest neighbors to be used}
}
\value{
A (m*p) matrix of predicted ratings for m testing users on all items
}
\description{
A function to predict ratings for testing users, given training user rating matrix.
}
\examples{
# Example code for UBCF
data("MovieLenseSub") # A subset of MovieLense Data
train <- MovieLenseSub[1:200,]
test <- MovieLenseSub[201:250,]
train[is.na(train)]<-0
test[is.na(test)]<-0

predicted<-ubcf_predict(train,test,k_neighbors=25)
topN.idx<-top_n_list(test,predicted,N=10)
}
